example/demo/requirements.txt
langgraph
langchain-core
langchain-openai
pyyaml
python-docx
